Fix potential memory leak in DefaultAdapterProvider
