+++ core/src/main/java/com/dooapp/fxform/adapter/FloatToDoubleAdapter.java
@@ -1,0 +1,14 @@
+package com.dooapp.fxform.adapter;
+
+public class FloatToDoubleAdapter implements Adapter<Float, Double> {
+
+    @Override
+    public Double adaptTo(Float from) {
+        return from.doubleValue();
+    }
+
+    @Override
+    public Float adaptFrom(Double to) {
+        return to.floatValue();
+    }
+}
