+++ core/src/main/java/com/dooapp/fxform/adapter/IntegerToDoubleAdapter.java
@@ -1,0 +1,14 @@
+package com.dooapp.fxform.adapter;
+
+public class IntegerToDoubleAdapter implements Adapter<Integer, Double> {
+
+    @Override
+    public Double adaptTo(Integer from) {
+        return from.doubleValue();
+    }
+
+    @Override
+    public Integer adaptFrom(Double to) {
+        return to.intValue();
+    }
+}
